Put a bunch of api state/data loader stuff into PropertyTable instead of the parent to resolve some React rendering complaints
